Separate transform_data and calculate_slices in transform_utils.py
